04-basic-servers/python/pyproject.toml
[tool.poetry]
name = "python"
version = "0.1.0"
description = ""
authors = ["Michael Fuglø <michael@fuglo.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.91.0"
uvicorn = "^0.20.0"
pyyaml = "^6.0"
requests = "^2.28.2"
python-dotenv = "^1.0.0"
xmltodict = "^0.13.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
